+++ screens/home/_components/index.ts
@@ -1,0 +1,3 @@
+export { TicketList } from './ticket-list';
+export { TicketCard } from './ticket-card';
+export { GenreFilter } from './genre-filter';
